create wallet router featured: CRUD and handling error

+++ controllers/category-controller.js
@@ -18,7 +18,6 @@ class Controller {
     try {
       const { name, type } = req.body;
       const response = await Category.create({ name, type });
-
       res.status(201).json({
         message: `Success create category with id ${response.id}, name ${response.name}, and type ${response.type}`,
       });
@@ -31,13 +30,17 @@ class Controller {
     try {
       const { name, type } = req.body;
       const id = req.params.categoryId;
+      if (isNaN(+id)) throw { name: "Invalid Id" };
       if (!name || !type) throw { name: "Invalid input" };
 
       const checkCategory = await Category.findByPk(id);
       if (!checkCategory) throw { name: "NotFound" };
 
-      const response = await Category.update({ name, type }, { where: { id } });
-      if (!response[0]) throw { name: "Invalid input" };
+      const updatedCategory = await Category.update(
+        { name, type },
+        { where: { id } }
+      );
+      if (!updatedCategory[0]) throw { name: "Invalid input" };
 
       res.status(200).json({
         message: `Success update category with id ${id}`,
@@ -50,6 +53,7 @@ class Controller {
   static async deleteCategory(req, res, next) {
     try {
       const { categoryId } = req.params;
+      if (isNaN(+categoryId)) throw { name: "Invalid Id" };
 
       const response = await Category.destroy({ where: { id: categoryId } });
       if (!response) throw { name: "NotFound" };
